Change list wording and update link to dev guide
I removed "Cloud Identity" from the second and third bullets, and instead put it at the end of the intro sentence. That change makes the bullet items more parallel. I also updated the link to the dev guide to go straight to the "Developer Guide" section of the doc page.